+++ src/generic-table/interfaces/gt-options.ts
@@ -1,0 +1,4 @@
+export interface GtOptions {
+  cache: boolean;
+  debounceTime: number;
+}