+++ src/generic-table/pipes/gt-chunk.pipe.ts
@@ -1,19 +1,23 @@
-import { Pipe, PipeTransform } from '@angular/core';
+import {Pipe, PipeTransform, EventEmitter} from '@angular/core';
+import {GtInformation} from '../interfaces/gt-information';
 
 @Pipe({
   name: 'gtChunk'
 })
 export class GtChunkPipe implements PipeTransform {
 
-  transform(array: any[], chunkSize: number, page: number, refreshPageArray: boolean, refreshData: number): any[] {
+  transform(array: any[],gtInfo:GtInformation, chunkSize: number, page: number, refreshPageArray: boolean, refreshData: number, gtEvent: EventEmitter<any>): any[] {
     //console.log('chunk array');
     //console.log(array,chunkSize,page);
     if(!Array.isArray(array)) return array;
-    var R = [];
-    for (var i=0,len=array.length; i<len; i+=chunkSize)
-      R.push(array.slice(i,i+chunkSize));
-    //console.log(R)
-    return R[page-1];
+    let pages = [];
+    for (let i=0,len=array.length; i<len; i+=chunkSize)
+      pages.push(array.slice(i,i+chunkSize));
+    gtInfo.pageTotal = pages.length;
+    setTimeout(()=>gtEvent.emit({
+      name:'gt-info',
+      value:gtInfo
+    }),0);
+    return pages[page-1];
   }
-
 }