+++ src/index.ts
@@ -1,5 +1,5 @@
 export * from './generic-table/components/gt-expanding-row.component';
-export * from './generic-table/generic-table.component';
+export * from './generic-table/components/generic-table.component';
 export * from './generic-table/generic-table.module';
 export * from './generic-table/interfaces/gt-config';
 export * from './generic-table/interfaces/gt-config-field';
@@ -8,3 +8,4 @@ export * from './generic-table/interfaces/gt-information';
 export * from './generic-table/interfaces/gt-paging-info';
 export * from './generic-table/interfaces/gt-row';
 export * from './generic-table/interfaces/gt-texts';
+export * from './generic-table/interfaces/gt-options';